+++ app/frontend/src/components/SimplySidebar/SimplySidebar.module.css
@@ -1,0 +1,3 @@
+.sidebar {
+    padding: var(--spacingVerticalS);
+}
