+++ app/frontend/src/components/SummarizeSidebar/SummarizeSidebar.module.css
@@ -1,0 +1,3 @@
+.sidebar {
+    padding: var(--spacingVerticalS);
+}
