--- app/frontend/src/pages/simplyfied-language/SimplySidebar.module.css
@@ -1,3 +1,0 @@
-.sidebar {
-    padding: var(-spacingS);
-}
