+++ app/frontend/src/pages/summarize/Summarize.tsx
@@ -15,6 +15,7 @@ import { LLMContext } from "../../components/LLMSelector/LLMContextProvider";
 import { ChatLayout } from "../../components/ChatLayout/ChatLayout";
 import { ChatTurnComponent } from "../../components/ChatTurnComponent/ChatTurnComponent";
 import { Sidebar } from "../../components/Sidebar/Sidebar";
+import { SummarizeSidebar } from "../../components/SummarizeSidebar/SummarizeSidebar";
 
 const STORAGE_KEY_LEVEL_OF_DETAIL = "SUM_LEVEL_OF_DETAIL";
 
@@ -92,28 +93,14 @@ const Summarize = () => {
 
     useEffect(() => chatMessageStreamEnd.current?.scrollIntoView({ behavior: "smooth" }), [isLoading]);
 
-    const onDetaillevelChanged = (e: any, selection: RadioGroupOnChangeData) => {
-        setDetaillevel(selection.value as "long" | "medium" | "short");
-        localStorage.setItem(STORAGE_KEY_LEVEL_OF_DETAIL, selection.value);
+    const onDetaillevelChanged = (newValue: string) => {
+        setDetaillevel(newValue as "long" | "medium" | "short");
+        localStorage.setItem(STORAGE_KEY_LEVEL_OF_DETAIL, newValue);
     };
 
     const examplesComponent = <ExampleListSum onExampleClicked={onExampleClicked} />;
-    const sidebar_actions = (
-        <>
-            <ClearChatButton onClick={clearChat} disabled={!lastQuestionRef.current || isLoading} />,
-        </>
-    );
-    const sidebar_content = (
-        <>
-            <Field label={t("sum.levelofdetail")}>
-                <RadioGroup layout="vertical" onChange={onDetaillevelChanged} value={detaillevel_pref}>
-                    <Radio value="short" label={t("sum.short")} />
-                    <Radio value="medium" label={t("sum.medium")} />
-                    <Radio value="long" label={t("sum.long")} />
-                </RadioGroup>
-            </Field>
-        </>
-    );
+    const sidebar_actions = <ClearChatButton onClick={clearChat} disabled={!lastQuestionRef.current || isLoading} />;
+    const sidebar_content = <SummarizeSidebar onDetaillevelChanged={onDetaillevelChanged} detaillevel_pref={detaillevel_pref} />;
     const sidebar = <Sidebar actions={sidebar_actions} content={sidebar_content}></Sidebar>;
     const answerList = (
         <>
